+++ RentAMovie/Models/MovieModuls/TmdbSingleMovieModel.cs
@@ -31,9 +31,9 @@
         [JsonProperty("backdrop_path")]
         public string? BackdropPath { get; set; }
 
-        public int? Budget { get; set; }
+        public long? Budget { get; set; }
 
-        public int? Revenue { get; set; }
+        public long? Revenue { get; set; }
 
         public int? Runtime { get; set; }
